# Conflicts:
#   src/main/resources/templates/user_management.html

+++ src/main/java/com/nsa/cubric/application/repositories/AccountRepository.java
@@ -66,7 +66,7 @@ public class AccountRepository implements AccountRepositoryStatic {
     }
 
     @Override
-    public boolean removeUser(Integer userId){
+    public boolean removeUser(Long userId){
 
         int rowsAffected = jdbcTemplate.update("DELETE FROM useraccounts WHERE id=?;",(userId));
         jdbcTemplate.update("DELETE FROM userprofiles WHERE useraccountid=?",(userId));
@@ -74,7 +74,7 @@ public class AccountRepository implements AccountRepositoryStatic {
     }
 
     @Override
-    public Integer removeUserResponses(Integer userId){
-        return jdbcTemplate.update("CALL removeUserRatings(?)", (userId));
+    public Long removeUserResponses(Long userId){
+        return Long.valueOf(jdbcTemplate.update("CALL removeUserRatings(?)", (userId)));
     }
 }
