+++ src/main/java/com/nsa/cubric/application/repositories/AccountRepositoryStatic.java
@@ -11,6 +11,6 @@ public interface AccountRepositoryStatic {
     public void insertNewAccount(AccountDTO account);
     public List<Account> getAll();
     public void insertNewProfile(ProfileDTO profile);
-    public boolean removeUser(Integer userId);
-    public Integer removeUserResponses(Integer userId);
+    public boolean removeUser(Long userId);
+    public Long removeUserResponses(Long userId);
 }
