+++ src/main/java/com/nsa/cubric/application/services/AdminServicesStatic.java
@@ -1,6 +1,7 @@
 package com.nsa.cubric.application.services;
 
 public interface AdminServicesStatic {
-    public boolean removeUser(Integer userId);
-    public Integer removeUserResponses(Integer userId);
+    public boolean removeUser(Long userId);
+    public Long removeUserResponses(Long userId);
+
 }
